RGB to HSV detection: Better color detection
Can now detect shades of color (Can play outside and in variable light situations)

Thanks Sam for tip

+++ app/src/main/java/com/taserlag/lasertag/player/Player.java
@@ -146,7 +146,7 @@ public class Player{
         dbPlayer.getPlayerStats().saveCaptain(captain, dbPlayerReference.child("playerStats"));
     }
 
-    public int[] getColor() {
+    public float[] getColor() {
         return dbUser.getColor();
     }
